requirements.txt
anyio==4.2.0
base58==2.1.1
cachetools==4.2.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jsonalias==0.1.1
pycparser==2.21
PyNaCl==1.5.0
python-telegram-bot==20.8
requests==2.31.0
rfc3986==1.5.0
setuptools==69.0.3
sniffio==1.3.0
solana==0.18.1
solders==0.20.0
types-cachetools==4.2.10
typing_extensions==4.9.0
urllib3==2.2.0
websockets==11.0.3
wheel==0.42.0
